requirements.txt
ultralytics
opencv-python
Pillow
easyocr
